refactor: set min. version of iOS to 12.0
