+++ Cookbook/Application/AppDelegate.swift
@@ -1,0 +1,24 @@
+//
+//  AppDelegate.swift
+//  Cookbook
+//
+//  Created by Егор Бадмаев on 15.10.2022.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+    
+    var window: UIWindow?
+    
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        
+        let window = UIWindow(frame: UIScreen.main.bounds)
+        window.rootViewController = ViewController()
+        window.makeKeyAndVisible()
+        self.window = window
+        
+        return true
+    }
+}
