+++ Cookbook/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
