+++ Cookbook/Resources/Assets.xcassets/Contents.json
